rename(fe): 전처리 스트로크 데이터 타입 파일명 카멜케이스로 수정
