+++ client/src/features/similarity/model/preprocessedStrokeData.ts
